refactor: organize admin ui files
