+++ src/ui/admin/components/ContributionBar.tsx
