+++ src/ui/admin/routes/languages.$code/exports.tsx
@@ -1,6 +1,6 @@
 import LoadingSpinner from "@/components/LoadingSpinner";
 import ViewTitle from "@/components/ViewTitle";
-import InterlinearExportPanel from "@/modules/export/ui/InterlinearExportPanel";
+import InterlinearExportPanel from "@/ui/admin/components/InterlinearExportPanel";
 import { createFileRoute } from "@tanstack/react-router";
 import { useTranslations } from "use-intl";
 import { Suspense } from "react";
